Actualizaciones y prueba del grafo

+++ project/src/motor_matching_engine/Usuario.java
@@ -2,10 +2,10 @@ package motor_matching_engine;
 
 public class Usuario {
     private int ID;
-    private long verticeIDOrigen;
-    private long verticeIDDestino;
+    private int verticeIDOrigen;
+    private int verticeIDDestino;
 
-    public Usuario(int ID,long verticeOrigen, long verticeDestino){
+    public Usuario(int ID,int verticeOrigen, int verticeDestino){
         this.ID = ID;
         this.verticeIDOrigen = verticeOrigen;
         this.verticeIDDestino = verticeDestino;
@@ -19,19 +19,19 @@ public class Usuario {
         ID = iD;
     }
 
-    public long getVerticeIDOrigen() {
+    public int getVerticeIDOrigen() {
         return verticeIDOrigen;
     }
 
-    public void setVerticeIDOrigen(long verticeIDOrigen) {
+    public void setVerticeIDOrigen(int verticeIDOrigen) {
         this.verticeIDOrigen = verticeIDOrigen;
     }
 
-    public long getVerticeIDDestino() {
+    public int getVerticeIDDestino() {
         return verticeIDDestino;
     }
 
-    public void setVerticeIDDestino(long verticeIDDestino) {
+    public void setVerticeIDDestino(int verticeIDDestino) {
         this.verticeIDDestino = verticeIDDestino;
     }
